requirements.txt
# Core dependencies
MDAnalysis>=2.0.0  # For molecular dynamics trajectory handling
numpy>=1.24.0      # Required by MDAnalysis
pandas>=2.0.0      # For data analysis and CSV handling
networkx>=3.0      # For molecular graph operations
scipy>=1.10.0      # For scientific computations
scikit-learn>=1.0.0  # For clustering
mdtraj>=1.9.0      # For trajectory analysis
tqdm>=4.65.0       # For progress bars
biopython>=1.81    # For PDB file handling
matplotlib>=3.7.0
rdkit>=2023.09.1

# Development dependencies
black>=23.0.0      # Code formatting
isort>=5.12.0      # Import sorting
pylint>=2.17.0     # Code linting
mypy>=1.0.0        # Type checking
pytest>=7.0.0      # Testing framework
pytest-cov>=4.0.0  # Test coverage reporting 